requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-Login>=0.5
Flask-Migrate>=3.0
Flask-WTF>=1.0  # For forms, CSRF protection
python-dotenv>=0.19 # For loading .env files
gunicorn>=20.0 # For production deployment
psycopg2-binary # If using PostgreSQL, otherwise can be removed or replaced by mysqlclient etc.
email_validator>=2.0.0 # Added for WTForms email validation
WTForms-SQLAlchemy>=0.3 # Added for QuerySelectField
pandas>=1.0 # For CSV import
# Add other dependencies as needed, e.g.:
# WeasyPrint # For PDF generation
